+++ xclim/indicators/atmos/_temperature.py
@@ -59,6 +59,7 @@ __all__ = [
     "warm_spell_duration_index",
     "maximum_consecutive_warm_days",
     "fire_season",
+    "corn_heat_units",
 ]
 
 
@@ -686,4 +687,16 @@ fire_season = Tasx(
     description="Fire season mask, computed with method {method}.",
     units="",
     compute=indices.fire_season,
+)
+
+corn_heat_units = TasminTasmax(
+    identifier="corn_heat_units",
+    units="",
+    long_name="Corn heat units (Tmin > {thresh_tasmin} and Tmax > {thresh_tasmax}).",
+    description="Temperature-based index used to estimate the development of corn crops. "
+    "Corn growth occurs when the minimum and maximum daily temperature both exceeds "
+    "specific thresholds : Tmin > {thresh_tasmin} and Tmax > {thresh_tasmax}.",
+    cell_methods="",
+    missing="skip",
+    compute=indices.corn_heat_units,
 )